fix: ci/cd integration tests
