+++ src/test/demo-sample.spec.js
@@ -1,9 +1,10 @@
 import { Selector } from "testcafe";
-import { validateIframeTexts, validateIssuerTexts, location } from "./helper";
+import { validateIframeTexts, validateIssuerTexts, navigateToVerify, location } from "./helper";
 
-fixture("Demo Sample Certificate Rendering").page`${location}`;
+fixture("Demo Sample Certificate Rendering").page`${location}/`;
 
 test("demo sample document is rendered correctly when dragged", async (t) => {
+  await navigateToVerify();
   await t.dragToElement(Selector("[draggable='true']"), Selector("#certificate-dropzone"));
   await validateIssuerTexts(["EXAMPLE.OPENATTESTATION.COM"]);
   await validateIframeTexts(["Name & Address of Shipping Agent/Freight Forwarder"]);
@@ -12,6 +13,7 @@ test("demo sample document is rendered correctly when dragged", async (t) => {
 // TODO, implement
 // eslint-disable-next-line jest/no-disabled-tests
 test.skip("demo sample document is rendered correctly when clicked", async (t) => {
+  await navigateToVerify();
   await t.resizeWindow(800, 600);
   await t.click(Selector("#demoClick"));
   await validateIssuerTexts(["EXAMPLE.OPENATTESTATION.COM"]);
