+++ src/test/did-verified.spec.js
@@ -2,7 +2,7 @@ import { uploadDocument, validateIframeTexts, validateIssuerTexts, location } fr
 
 fixture("DID Certificate Rendering").page`${location}`;
 
-test("sample document is rendered correctly when dns is verified", async () => {
+test("sample document is rendered correctly when dns is verified", async (t) => {
   await uploadDocument("./fixture/sample-did-verified.json");
   await validateIssuerTexts(["ETHR:0XE712878F6E8D5D4F9E87E10DA604F9CB564C9A89"]);
