+++ src/test/ebl-accepted-surrender.spec.js
@@ -1,5 +1,5 @@
 import { Selector } from "testcafe";
-import { location } from "./helper";
+import { navigateToVerify, location } from "./helper";
 
 fixture("Render red surrender sign if ebl is accepted surrendered").page`${location}`;
 
@@ -10,6 +10,7 @@ const SurrenderedSign = Selector("#surrendered-sign");
 const InteractionAvaliable = Selector("[data-testid='connectToWallet']");
 
 test("should displays surrendered sign when document is owned by 0xdead address", async (t) => {
+  await navigateToVerify();
   const container = Selector("#certificate-dropzone");
   await container();
   await t.setFilesToUpload("input[type=file]", [Document]);
