+++ src/test/ebl-surrendered.spec.js
@@ -1,5 +1,5 @@
 import { Selector } from "testcafe";
-import { location } from "./helper";
+import { navigateToVerify, location } from "./helper";
 
 fixture("Render red surrender sign if ebl is surrendered").page`${location}`;
 
@@ -11,6 +11,7 @@ const SurrenderedSign = Selector("#surrender-sign");
 const InteractionAvaliable = Selector("[data-testid='connectToWallet']");
 
 test("should displays surrender sign when document is owned by token registry", async (t) => {
+  await navigateToVerify();
   const container = Selector("#certificate-dropzone");
   await container();
   await t.setFilesToUpload("input[type=file]", [SurrenderedDocument]);
@@ -23,6 +24,7 @@ test("should displays surrender sign when document is owned by token registry", 
 });
 
 test("should not displays surrender sign when document is not owned by token registry", async (t) => {
+  await navigateToVerify();
   const container = Selector("#certificate-dropzone");
   await container();
   await t.setFilesToUpload("input[type=file]", [NonSurrenderedDocument]);
