+++ src/test/ebl-wallet-owner.spec.js
@@ -1,5 +1,5 @@
 import { Selector } from "testcafe";
-import { location } from "./helper";
+import { navigateToVerify, location } from "./helper";
 
 fixture("Render document for direct ownership").page`${location}`;
 
@@ -9,6 +9,7 @@ const DocumentStatus = Selector("#document-status");
 const InteractionAvaliable = Selector("[data-testid='connectToWallet']");
 
 test("Displays action button for wallet owners", async (t) => {
+  await navigateToVerify();
   const container = Selector("#certificate-dropzone");
   await container();
   await t.setFilesToUpload("input[type=file]", [Document]);
