+++ src/test/helper.js
@@ -5,6 +5,7 @@ export const SampleTemplate = Selector("#root");
 const CertificateDropzone = Selector("#certificate-dropzone");
 const DocumentStatus = Selector("#document-status");
 const IssuedByDomainName = Selector("#issuedby .domain");
+const VerifyPage = Selector("a[href='/verify']");
 
 export const validateTextContent = async (testcafe, component, texts) =>
   texts.reduce(
@@ -12,7 +13,11 @@ export const validateTextContent = async (testcafe, component, texts) =>
     Promise.resolve()
   );
 
+export const navigateToVerify = async () => {
+  await t.click(VerifyPage);
+};
 export const uploadDocument = async (documentPath) => {
+  await navigateToVerify();
   await CertificateDropzone();
   await t.setFilesToUpload("input[type=file]", [documentPath]);
 };
@@ -30,4 +35,4 @@ export const validateIssuerTexts = async (texts) => {
 
 export const CloseWindow = ClientFunction(() => window.close());
 
-export const location = "http://localhost:3000/verify";
+export const location = "http://localhost:3000";
